RealTime: changes for AVR compatibility
git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@2063 235a714a-1e17-0410-b3bb-947d69a96898

+++ include/mw/el/ml/SerializationBase.h
@@ -100,7 +100,7 @@ namespace famouso {
                         }
 
                         bool further_space(uint16_t bytes) const {
-                            return data_end - data.u8 >= bytes;
+                            return (uint16_t)(data_end - data.u8) >= bytes;
                         }
                         bool buffer_full() const {
                             return data_end  == data.u8;
